tests are placed in separate project
