+++ cs/MarkdownTests/MarkdownSpec.md
@@ -1,0 +1,73 @@
+# Спецификация языка разметки
+
+Посмотрите этот файл в сыром виде. Сравните с тем, что показывает github.
+Все совпадения случайны ;)
+
+
+
+# Курсив
+
+Текст, _окруженный с двух сторон_ одинарными символами подчерка,
+должен помещаться в HTML-тег \<еm> вот так:
+
+Текст, \<еm>окруженный с двух сторон\</еm> одинарными символами подчерка,
+должен помещаться в HTML-тег \<еm>.
+
+
+
+# Полужирный
+
+__Выделенный двумя символами текст__ должен становиться полужирным с помощью тега \<strоng>.
+
+
+
+# Экранирование
+
+Любой символ можно экранировать, чтобы он не считался частью разметки.
+\_Вот это\_, не должно выделиться тегом \<еm>.
+
+Символ экранирования исчезает из результата, только если экранирует что-то.
+Здесь сим\волы экранирования\ \должны остаться.\
+
+Символ экранирования тоже можно экранировать: \\_вот это будет выделено тегом_ \<еm>
+
+
+
+# Взаимодействие тегов
+
+Внутри __двойного выделения _одинарное_ тоже__ работает.
+
+Но не наоборот — внутри _одинарного __двойное__ не_ работает.
+
+Подчерки внутри текста c цифрами_12_3 не считаются выделением и должны оставаться символами подчерка.
+
+Однако выделять часть слова они могут: и в _нач_але, и в сер_еди_не, и в кон_це._
+
+В то же время выделение в ра_зных сл_овах не работает.
+
+__Непарные_ символы в рамках одного абзаца не считаются выделением.
+
+За подчерками, начинающими выделение, должен следовать непробельный символ. Иначе эти_ подчерки_ не считаются выделением 
+и остаются просто символами подчерка.
+
+Подчерки, заканчивающие выделение, должны следовать за непробельным символом. Иначе эти _подчерки _не считаются_ окончанием выделения 
+и остаются просто символами подчерка.
+
+В случае __пересечения _двойных__ и одинарных_ подчерков ни один из них не считается выделением.
+
+Если внутри подчерков пустая строка ____, то они остаются символами подчерка.
+
+
+
+# Заголовки
+
+Абзац, начинающийся с "# ", выделяется тегом \<h1> в заголовок.
+В тексте заголовка могут присутствовать все прочие символы разметки с указанными правилами.
+
+Таким образом
+
+# Заголовок __с _разными_ символами__
+
+превратится в:
+
+\<h1>Заголовок \<strоng>с \<еm>разными\</еm> символами\</strоng>\</h1>
