+++ cs/MarkdownTests/MdTests.cs
@@ -1,8 +1,9 @@
 ﻿using System.IO;
+using Markdown;
 using Markdown.Renderers;
 using NUnit.Framework;
 
-namespace Markdown
+namespace MarkdownTests
 {
     class MdTests
     {
@@ -23,8 +24,7 @@ namespace Markdown
         public string MdRender_ConvertItalicText_When(string markdownText) =>
             MdRender(markdownText);
 
-
-        [TestCase("внутри текста c цифрами_12_3 не считаются выделением", ExpectedResult = "внутри текста c цифрами_12_3 не считаются выделением", TestName = "inside text with numbers")]
+        [TestCase("witn numbers_12_3 text", ExpectedResult = "witn numbers_12_3 text", TestName = "inside text with numbers")]
         [TestCase("_text___", ExpectedResult = "_text___", TestName = "blink string")]
         [TestCase("text_ text_", ExpectedResult = "text_ text_", TestName = "after each tag space")]
         [TestCase("_text text _text", ExpectedResult = "_text text _text", TestName = "end tag before symbol")]
@@ -34,36 +34,31 @@ namespace Markdown
         public string MdRender_NotConvertItalicText_When(string markdownText) =>
             MdRender(markdownText);
 
-
         [TestCase("__text__", ExpectedResult = "<strong>text</strong>", TestName = "text between tags")]
         public string MdRender_ConvertBoldText_When(string markdownText) =>
             MdRender(markdownText);
 
         [TestCase(@"\_text\_", ExpectedResult = @"_text_", TestName = "single comment")]
+        [TestCase(@"\\\_text\\\_", ExpectedResult = @"\_text\_", TestName = "triple comment")]
         public string MdRender_ConvertCommentedTags_When(string markdownText) =>
             MdRender(markdownText);
 
         [TestCase(@"\\_text\\_", ExpectedResult = @"\<em>text\</em>", TestName = "double comment")]
+        [TestCase(@"\\\\_text\\\\_", ExpectedResult = @"\\<em>text\\</em>", TestName = "quad comment")]
         public string MdRender_NotConvertCommentedTags_When(string markdownText) =>
             MdRender(markdownText);
 
-        [TestCase("__t_t_t_t__", ExpectedResult = "<strong>t<em>t</em>t_t</strong>", TestName = "Text2")]
-        [TestCase("__t_t___t__", ExpectedResult = "<strong>t_t___t</strong>", TestName = "Text3")]
-        [TestCase(@"\__t_t___t__", ExpectedResult = @"__t_t___t__", TestName = "Text4")]
+        [TestCase("__t_t_t_t__", ExpectedResult = "<strong>t<em>t</em>t_t</strong>", TestName = "Italic inside bold into word")]
+        [TestCase("__t_t___t__", ExpectedResult = "<strong>t_t___t</strong>", TestName = "Italic with blink end inside bold into word")]
+        [TestCase(@"\__t_t___t__", ExpectedResult = @"__t_t___t__", TestName = "commented start tag")]
         [TestCase("t _t __t__ t_ t", ExpectedResult = @"t <em>t __t__ t</em> t", TestName = "bold inside italic")]
         [TestCase("t __t _t_ t__ t", ExpectedResult = @"t <strong>t <em>t</em> t</strong> t", TestName = "Italic inside bold")]
         [TestCase("t _t t __t t_ t__ t", ExpectedResult = "t _t t __t t_ t__ t", TestName = "one tag start before end other and end after")]
-        [TestCase("__t _t t__ t t_ _t t__ t t_ t__ t", ExpectedResult = "__t _t t__ t t_ <em>t t__ t t</em> t__ t", TestName = "mix2")]
-        [TestCase("__t _t t__ t t_ _t t__ t t_ t t", ExpectedResult = "__t _t t__ t t_ <em>t t__ t t</em> t t", TestName = "mix")]
-        [TestCase("_t __t t_ t t__ _t __t t_ t t__ t", ExpectedResult = "_t __t t_ t t__ _t __t t_ t t__ t", TestName = "mix3")]
-        [TestCase("_t __t t_ t t__ __t t_ t t__ t t", ExpectedResult = "_t __t t_ t t__ <strong>t t_ t t</strong> t t", TestName = "mix4")]
-        [TestCase(@"Здесь сим\волы экранирования\ \должны остаться.\", ExpectedResult = @"Здесь сим\волы экранирования\ \должны остаться.\", TestName = "Comment 2")]
-        [TestCase(@"\_Вот это\_, не должно выделиться тегом <em>", ExpectedResult = @"_Вот это_, не должно выделиться тегом <em>", TestName = "Comment 1")]
+        [TestCase("__t _t t__ t t_ _t t__ t t_ t__ t", ExpectedResult = "__t _t t__ t t_ <em>t t__ t t</em> t__ t", TestName = "intersections that start with bold")]
+        [TestCase("_t __t t_ t t__ _t __t t_ t t__ t", ExpectedResult = "_t __t t_ t t__ _t __t t_ t t__ t", TestName = "intersections that start with italic")]
+        [TestCase(@"Здесь сим\волы экранирования\ \должны остаться.\", ExpectedResult = @"Здесь сим\волы экранирования\ \должны остаться.\", TestName = "comment without tag")]
         public string MdRender_ConvertTextWithDifferentTags_When(string markdownText) =>
             MdRender(markdownText);
-
-
-
 
         [TestCase("         # text", ExpectedResult = "         <h1>text</h1>", TestName = "header tag is not at the beginning of the text")]
         [TestCase("# text", ExpectedResult = "<h1>text</h1>", TestName = "header tag is at the beginning of the text")]
@@ -76,8 +71,8 @@ namespace Markdown
         public string MdRender_NotConvertHeaderText_When(string markdownText) =>
             MdRender(markdownText);
 
-        [TestCase(@"..\..\..\..\..\Markdown.md", TestName = "Markdown")]
-        [TestCase(@"..\..\..\..\..\MarkdownSpec.md", TestName = "MarkdownSpec")]
+        [TestCase("Markdown.md", TestName = "Markdown")]
+        [TestCase("MarkdownSpec.md", TestName = "MarkdownSpec")]
         public void MdRender_MdFileToHtml_CreatesFor(string path)
         {
             var mdSpec = File.ReadAllText(path);
